# [GrCUDA-96-8] Better device policy selection & mocked benchmarks & fixes to python benchmark suite