--- a/projects/com.nvidia.grcuda/src/com/nvidia/grcuda/GrCUDAOptions.java
+++ b/projects/com.nvidia.grcuda/src/com/nvidia/grcuda/GrCUDAOptions.java
@@ -97,4 +97,8 @@ public final class GrCUDAOptions {
 
     @Option(category = OptionCategory.USER, help = "Set the location of the CSV file that contains the estimated bandwidth between each CPU and GPU in the system.", stability = OptionStability.EXPERIMENTAL) //
     public static final OptionKey<String> BandwidthMatrix = new OptionKey<>(GrCUDAOptionMap.DEFAULT_BANDWIDTH_MATRIX);
+
+    @Option(category = OptionCategory.USER, help = "When selecting a device, do not give priority to devices that have less than this percentage of data already available, " +
+            "in some DeviceSelectionPolicies such as min-transfer-size. A lower percentage favors exploitation, a high percentage favors exploration.", stability = OptionStability.EXPERIMENTAL) //
+    public static final OptionKey<Double> DataThreshold = new OptionKey<>(GrCUDAOptionMap.DEFAULT_DATA_THRESHOLD);
 }