--- a/projects/com.nvidia.grcuda/src/com/nvidia/grcuda/runtime/CUDARuntime.java
+++ b/projects/com.nvidia.grcuda/src/com/nvidia/grcuda/runtime/CUDARuntime.java
@@ -115,7 +115,7 @@ public final class CUDARuntime {
      * Identifier of the GPU that is currently active;
      */
     private int currentGPU = DEFAULT_DEVICE;
-
+    
     public boolean isMultiGPUEnabled() {
         return this.numberOfGPUsToUse > 1;
     }