--- a/projects/com.nvidia.grcuda/src/com/nvidia/grcuda/runtime/stream/policy/DeviceSelectionPolicyEnum.java
+++ b/projects/com.nvidia.grcuda/src/com/nvidia/grcuda/runtime/stream/policy/DeviceSelectionPolicyEnum.java
@@ -14,7 +14,8 @@ public enum DeviceSelectionPolicyEnum {
         this.name = name;
     }
 
-    public final String getName() {
+    @Override
+    public String toString() {
         return name;
     }
 }