--- a/projects/resources/connection_graph/run.sh
+++ b/projects/resources/connection_graph/run.sh
@@ -1,3 +1,4 @@
 #!/bin/sh
 make
+mkdir -p datasets;
 bin/connection_graph