* **Added mocked benchmarks:** for each multi-gpu benchmark in our suite, there is a mocked version where we check that the GPU assignment is the one we expect. Added utility functions to easily test mocked benchmarks
* simplified round-robin device selection policy, now it works more or less as before but it is faster to update when using a subset of devices
* Added threshold parameter for data-aware device selection policies. When using `min-transfer-size` or `minmax/min-transfer-time`, consider only devices that have at least 10% (or X %) of the requested data. Basically, if a device only has a very small amount of data already available it is not worth preferring it to other devices, and it can cause scheduling to converge to a unique device. See B9 and B11, for example.
* Updated python benchmark suite to use new options, and optimized initialization of B1 (it is faster now) and B9 (it didn't work on matrices with 50K rows, as python uses 32-bit array indexing)
* Fixed performance regression in DeviceArray access. For a simple python code that writes 160M values on a DeviceArray, performance went from 4sec to 20sec by using GraalVM 21.3 instead of 21.2. Reverted GraalVM to 21.2. Using non-static final Logger in GrCUDAComputationalElement increased time from 4sec to 130sec (not sure why, they are not created in repeated array accesses): fixed this regression.

**Do not merge: PR stacked on top of [GRCUDA-96-7], to be merged on [GRCUDA-96-0] after [GRCUDA-96-7]**